Add files via upload
This is the original code structure that I used - it is messy and contains multiple trials of different types of modeling. I kept this as is to show different ways of thinking about the project and how I came to the end code conclusions that I did. This document is purely for reference for myself or for exploration if desired.